cli: Move shared file argument handling
The CLI parser uses the same --file and --files option shape across
commands that work on live or batch files.

The parser owns both option registration and argparse storage
normalization, so every reader has to step through shared file argument
mechanics before reaching command-specific parser branches.

This commit addresses that by adding a file_arguments module for shared
file option registration and parsed namespace normalization, then routing
parser construction through that module.

The following tests commit will cover the normalization helper and pin the
parser boundary for shared file arguments.
